+++ src/main/java/leets/leenk/domain/birthday/application/dto/request/BirthdayLetterRequest.java
@@ -1,4 +1,4 @@
-package leets.leenk.domain.birthday.application.dto;
+package leets.leenk.domain.birthday.application.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
