+++ src/main/java/leets/leenk/domain/birthday/application/usecase/BirthdayLetterUseCase.java
@@ -1,6 +1,6 @@
 package leets.leenk.domain.birthday.application.usecase;
 
-import leets.leenk.domain.birthday.application.dto.BirthdayLetterRequest;
+import leets.leenk.domain.birthday.application.dto.request.BirthdayLetterRequest;
 import leets.leenk.domain.birthday.application.dto.response.MyBirthdayLettersResponse;
 import leets.leenk.domain.birthday.application.exception.NotBirthdayTodayException;
 import leets.leenk.domain.birthday.application.mapper.BirthdayLetterMapper;
